fixing negative event numbers. Details: https://its.cern.ch/jira/browse/ATLASVPONE-287. (VP1Gui-00-08-11)
2016-05-31 Riccardo.Maria.Bianchi@cern.ch
    * fixing negative event numbers.
	  Details: https://its.cern.ch/jira/browse/ATLASVPONE-287
	* new tag: VP1Gui-00-08-11

2016-05-19 Riccardo.Maria.Bianchi@cern.ch
    * replaced in CMakeLists.txt 'atlas_install_scripts' with 'atlas_install_runtime' to fix teh
	  installation of the ATLAS logo PNG file in the share/ folder.
    * new tag: VP1Gui-00-08-10

2016-05-18 Riccardo.Maria.Bianchi@cern.ch
    * saving the name of the latest image produced in a cache file, for the
	batch production. More details here:
	https://its.cern.ch/jira/browse/ATLASVPONE-220
    * tag: VP1Gui-00-08-09

2016-03-01 Attila Krasznahorkay <Attila.Krasznahorkay@cern.ch>
	* Copied the UI files from the header directory back to src/.
	  As the CMT build can't pick them up from the header directory.
...
(Long ChangeLog diff - truncated)


Former-commit-id: 23f2902c951af97aff0b41493bc31d442bd01615

+++ graphics/VP1/VP1Gui/VP1Gui/VP1EventFile.h
@@ -25,7 +25,7 @@ public:
   //The interesting bits:
   //Event identity:
   int runNumber() const;
-  int eventNumber() const;
+  unsigned long long eventNumber() const;
   unsigned rawTime() const;
   QDateTime time() const;
 
@@ -45,7 +45,7 @@ public:
   VP1EventFile(const VP1EventFile & );
   VP1EventFile & operator= (const VP1EventFile & );
   VP1EventFile(const QString& filename,const QString& md5sum,
-	       int runnumber,int eventnumber,unsigned time, bool isvalid = true );
+	       int runnumber, unsigned long long eventnumber,unsigned time, bool isvalid = true );
   VP1EventFile(const QString&, const QString& md5sum = 0);//will attempt to decode from file name (e.g. vp1_234_123_1321334.pool.root). Instance becomes invalid if not succesful.
 private:
   class Imp;
