+++ graphics/VP1/VP1Gui/src/VP1EventDisplayEditor.cxx
@@ -15,7 +15,7 @@
 ////////////////////////////////////////////////////////////////
 
 #include "VP1Gui/VP1EventDisplayEditor.h"
-#include "ui_vp1eventdisplaywidget.h"
+#include "ui_vp1eventdisplayeditor.h"
 
 #include "VP1Base/VP1ExaminerViewer.h"
 #include "VP1Base/VP1QtInventorUtils.h"
@@ -139,7 +139,7 @@ public:
  * new version
  */
 //____________________________________________________________________
-VP1EventDisplayEditor::VP1EventDisplayEditor(VP1MainWindow* mainWindow, QList<int> listRunEventNumberTimestamp)
+VP1EventDisplayEditor::VP1EventDisplayEditor(VP1MainWindow* mainWindow, QList<unsigned long long> listRunEventNumberTimestamp)
 //: QWidget(0,Qt::WindowStaysOnTopHint), d(new Imp)
 : QWidget(0), d(new Imp)
 {
